requirements.txt
Django==1.9.6
wsgiref==0.1.2
djangorestframework==3.3.2
MySQL-python==1.2.5
wheel==0.24.0

